requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
requests==2.31.0
bleach==6.0.0
gunicorn==21.2.0
pyjwt==2.8.0
